requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==2.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
certifi>=2024.2.2
dnspython>=2.3.0
pymongo>=4.7.2
gunicorn==21.2.0
openpyxl>=3.1.2
requests>=2.31.0
